+++ src/components/commons/comment/commentItem/CommentItem.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/nextjs-vite";
 import CommentItem from "./CommentItem";
-import type { IBoardComment } from "@/shared/graphql/generated/types";
+import type { IBoardComment } from "@/api/graphql/generated/types";
 import { CommentActionsProvider } from "../context/CommentActionsContext";
 
 const mockActions = {
