+++ src/components/commons/profile/Profile.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/nextjs-vite";
-import type { IBoard } from "@/shared/graphql/generated/types";
+import type { IBoard } from "@/api/graphql/generated/types";
 import Profile from "./Profile";
 
 const IMG = "https://picsum.photos/id/64/200/200";
