+++ src/components/features/auth/sign-up/hooks/useCreateUser.ts
@@ -1,9 +1,9 @@
-import { IS_NEW_API } from "@/lib/config";
+import { IS_NEW_API } from "@/api/config";
 import { useToast } from "@/components/commons/toast/ToastProvider";
 import {
   IMutation,
   IMutationCreateUserArgs,
-} from "@/shared/graphql/generated/types";
+} from "@/api/graphql/generated/types";
 import { useNavigation } from "@/shared/hooks/ui/useNavigation";
 import { gql, useMutation } from "@apollo/client";
