+++ src/components/features/home/hooks/queries/useFetchBestRecords.ts
@@ -1,5 +1,5 @@
-import { IS_NEW_API } from "@/lib/config";
-import { IQuery } from "@/shared/graphql/generated/types";
+import { IS_NEW_API } from "@/api/config";
+import { IQuery } from "@/api/graphql/generated/types";
 import { gql, useQuery } from "@apollo/client";
 
 const idField = IS_NEW_API ? "id" : "_id";
