+++ src/components/features/record-comments/hooks/mutations/useCreateRecordComment.tsx
@@ -1,9 +1,9 @@
-import { IS_NEW_API } from "@/lib/config";
+import { IS_NEW_API } from "@/api/config";
 import { useToast } from "@/components/commons/toast/ToastProvider";
 import {
   IMutation,
   IMutationCreateBoardCommentArgs,
-} from "@/shared/graphql/generated/types";
+} from "@/api/graphql/generated/types";
 import { gql, useMutation } from "@apollo/client";
 
 const idField = IS_NEW_API ? "id" : "_id";
