+++ src/components/features/record-comments/hooks/queries/useFetchRecordComments.ts
@@ -1,8 +1,8 @@
-import { IS_NEW_API } from "@/lib/config";
+import { IS_NEW_API } from "@/api/config";
 import {
   IQuery,
   IQueryFetchBoardCommentsArgs,
-} from "@/shared/graphql/generated/types";
+} from "@/api/graphql/generated/types";
 import { gql, useQuery } from "@apollo/client";
 import { useState } from "react";
