+++ src/components/features/record/hooks/useFetchRecord.ts
@@ -1,5 +1,5 @@
-import { IS_NEW_API } from "@/lib/config";
-import { IQuery, IQueryFetchBoardArgs } from "@/shared/graphql/generated/types";
+import { IS_NEW_API } from "@/api/config";
+import { IQuery, IQueryFetchBoardArgs } from "@/api/graphql/generated/types";
 import { gql, useQuery } from "@apollo/client";
 
 const idField = IS_NEW_API ? "id" : "_id";
