+++ src/components/features/record/list/RecordCard/RecordCardBottom/RecordCardBottom.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/nextjs-vite";
 import RecordCardBottom from "./RecordCardBottom";
-import type { IBoard } from "@/shared/graphql/generated/types";
+import type { IBoard } from "@/api/graphql/generated/types";
 
 const IMG = "https://picsum.photos/id/64/200/200";
