+++ src/components/features/record/list/RecordCard/RecordFeedCard.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/nextjs-vite";
-import type { IBoard } from "@/shared/graphql/generated/types";
+import type { IBoard } from "@/api/graphql/generated/types";
 import RecordCard from "./RecordCard";
 
 const IMG = "https://picsum.photos/id/64/600/800";
