+++ src/shared/stores/user.ts
@@ -1,4 +1,4 @@
-import { IQuery } from "@/shared/graphql/generated/types";
+import { IQuery } from "@/api/graphql/generated/types";
 import { atom } from "recoil";
 
 export type LoggedInUser = IQuery["fetchUserLoggedIn"] | null;
